refactor: 패키지 구조 수정
- 최상위 main 패키지를 사다리게임의 패키지라는 의미로 ladder 패키지로 이름 변경
- 입력을 담당하는 클래스들을 view라는 패키지로 이동
  - InputController를 InputView로 이름 변경
- domain 패키지에 사다리 게임과 관련된 클래스들을 담았습니다.
