+++ src/ladder/domain/Ladder.java
@@ -1,4 +1,4 @@
-package main.laddergame;
+package ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
