+++ src/ladder/domain/LadderGame.java
@@ -1,6 +1,5 @@
-package main.laddergame;
+package ladder.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class LadderGame {
